+++ src/main/scala/org/podval/tools/nonjvm/RunTask.scala
@@ -3,11 +3,14 @@ package org.podval.tools.nonjvm
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.TaskAction
 import org.podval.tools.build.{BackendTask, TestEnvironment}
-import org.podval.tools.gradle.{TaskWithDependency, TaskWithRunner}
+import org.podval.tools.gradle.{TaskWithDependency, TaskWithOutput, TaskWithRunner}
 import org.podval.tools.test.task.TestTask
 import scala.reflect.ClassTag
 
-trait RunTask[B <: NonJvmBackend, L <: LinkTask[B] : ClassTag] extends BackendTask[B] with TaskWithDependency[L]:
+trait RunTask[B <: NonJvmBackend, L <: LinkTask[B] : ClassTag] extends BackendTask[B]
+  with TaskWithDependency[L]
+  with TaskWithOutput:
+  
   final protected def linkTask: L = dependency
 
   protected def run: Run[B]