+++ src/main/scala/org/podval/tools/scalajs/ScalaJSBuild.scala
@@ -1,21 +1,22 @@
 package org.podval.tools.scalajs
 
+import org.gradle.api.logging.LogLevel
 import org.podval.tools.nonjvm.Build
+import org.podval.tools.platform.Output
 import org.scalajs.logging.{Level as LevelJS, Logger as LoggerJS}
-import org.slf4j.event.Level
 
-class ScalaJSBuild(logSource: String) extends Build[LoggerJS](logSource):
+class ScalaJSBuild(output: Output) extends Build[LoggerJS](output):
   final override protected def interceptedExceptions: Set[Class[? <: Exception]] = Set.empty
 
   final override protected def backendLogger: LoggerJS = new LoggerJS:
-    override def trace(t: => Throwable): Unit = logThrowable(t)
+    override def trace(throwable: => Throwable): Unit = logThrowable(throwable)
 
     given CanEqual[LevelJS, LevelJS] = CanEqual.derived
-    override def log(level: LevelJS, message: => String): Unit = logAtLevel(
-      message,
-      level match
-        case LevelJS.Error => Level.ERROR
-        case LevelJS.Warn  => Level.WARN
-        case LevelJS.Info  => Level.INFO
-        case LevelJS.Debug => Level.DEBUG
+    override def log(logLevel: LevelJS, message: => String): Unit = logAtLevel(
+      message = message,
+      logLevel = logLevel match
+        case LevelJS.Error => LogLevel.ERROR
+        case LevelJS.Warn  => LogLevel.WARN
+        case LevelJS.Info  => LogLevel.INFO
+        case LevelJS.Debug => LogLevel.DEBUG
     )