+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.4] - 2025-07-28
+- feat: `jsLink.esVersion`;
+- chore: Gradle 9.0.0-rc-4;
+- chore: dependency updates;
+- cleanup;
+- 
 ## [0.9.3] - 2025-07-21
 - feat: `ZIO Test` on `Scala.js` and `Scala Native`;
 - chore: dependency updates;
@@ -378,14 +384,14 @@ ScalaCheck *does* produce nested tasks - adjusted TestClassProcessor to handle t
 - refactor: use sbt Selectors in filtering tests
 - wip: filtering suites based on tags
 - test: move test projects into `test-projects`
-- chore: latest Gradle, Zinc, opentorah, ScalaTest etc.
+- chore: latest Gradle, Zinc, OpenTorah, ScalaTest etc.
 
 ## [0.1.0] - 2022-07-18
 - testing integrated with Gradle;
 - testing integrated with IntelliJ Idea;
 - running sbt test frameworks without ScalaJS;
 - working on test filtering;
-- latest Gradle, Zinc and opentorah;
+- latest Gradle, Zinc and OpenTorah;
 
 **BREAKING CHANGES:**
 - no more extension;